+++ libs/data/src/lib/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `e1d823ea607e0a13193d18ced129faa041d2e9b3`;
+export const extractsHash = `3cc5522c9b1f96dd44e6567d5ac5e1142213f511`;